requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
pylint==2.6.0
pytest==6.1.1
requests==2.20.1
gunicorn==20.0.4
psycopg2==2.8.4
Flask-Mail==0.9.1
python-dotenv==0.15.0
zappa==0.52.0
Flask-S3==0.3.3
